+++ CHANGELOG.md
@@ -2,25 +2,71 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## [2.3.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v1.2.0...v2.3.0) (2022-09-06)
+## [2.7.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v1.2.0...v2.7.0) (2022-02-23)
 
 
 ### Features
 
-* **npm:** publish npm ([afc0e9a](https://github.com/TropixInc/pixchain-react-metamask/commit/afc0e9a26350e36218f35028c5ad35a4529789fd))
+* export wallet connect/disconnect function ([d16667e](https://github.com/TropixInc/pixchain-react-metamask/commit/d16667ebea8dabfe9f5e85c40559054f3ea02f91))
 * suport different hooks on signing ([6e65e1e](https://github.com/TropixInc/pixchain-react-metamask/commit/6e65e1e9b4bec8b3343142b782e5112bd9c2c36b))
 
 ## [1.1.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v1.0.3...v1.1.0) (2022-01-27)
 
-## [2.2.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v1.2.0...v2.2.0) (2022-09-06)
+### [2.6.2](https://github.com/TropixInc/pixchain-react-metamask/compare/v2.6.1...v2.6.2) (2022-02-22)
+
+
+### Bug Fixes
+
+* add dependency to hook ([2a723db](https://github.com/TropixInc/pixchain-react-metamask/commit/2a723db3b30055c5c32d4aa3cc9b7161a7faa403))
+
+### [2.6.1](https://github.com/TropixInc/pixchain-react-metamask/compare/v2.6.0...v2.6.1) (2022-02-22)
+
+
+### Bug Fixes
+
+* copy debounce code instead of install package ([cfe74f5](https://github.com/TropixInc/pixchain-react-metamask/commit/cfe74f58c543a39e57c1f320f52fc5798132af29))
+
+## [2.6.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v2.5.0...v2.6.0) (2022-02-22)
 
 
 ### Features
 
-* **npm:** publish npm ([afc0e9a](https://github.com/TropixInc/pixchain-react-metamask/commit/afc0e9a26350e36218f35028c5ad35a4529789fd))
+* add debounce to metamask signature request ([c458c87](https://github.com/TropixInc/pixchain-react-metamask/commit/c458c87f53ff0ffa4f3e8335fbb4026871ee9963))
+
+## [2.5.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v2.4.0...v2.5.0) (2022-02-21)
+
+
+### Features
+
+* add processing flag ([f5cd088](https://github.com/TropixInc/pixchain-react-metamask/commit/f5cd088622c2b6a884473866ecedfe252335a64a))
+
+
+### Bug Fixes
+
+* merge issue ([078fce1](https://github.com/TropixInc/pixchain-react-metamask/commit/078fce1ac3da5119a64190f0666d02a00c0f6c69))
+
+## [2.4.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v2.3.0...v2.4.0) (2022-02-21)
+
+
+### Features
+
+* add on rejection hook ([8772c47](https://github.com/TropixInc/pixchain-react-metamask/commit/8772c4733095fab7ae6b82807181363fbfc2a94c))
+
+## [2.3.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v1.2.0...v2.3.0) (2022-02-21)
+
+
+### Features
+
 * suport different hooks on signing ([6e65e1e](https://github.com/TropixInc/pixchain-react-metamask/commit/6e65e1e9b4bec8b3343142b782e5112bd9c2c36b))
 
 ## [1.1.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v1.0.3...v1.1.0) (2022-01-27)
+
+## [2.2.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v2.1.0...v2.2.0) (2022-02-18)
+
+
+### Features
+
+* export add transaction ([619aa4a](https://github.com/TropixInc/pixchain-react-metamask/commit/619aa4aa631680d120753000eecd3a9573bde657))
 
 ## [2.1.0](https://github.com/TropixInc/pixchain-react-metamask/compare/v1.2.0...v2.1.0) (2022-02-18)